requirements.txt
fastapi
uvicorn
pdfplumber
faiss-cpu
psycopg2-binary
transformers
openai
sentence-transformers
python-multipart
pydantic
numpy
pandas
pillow
requests
tqdm
flask
